rename project and associated files to assessments-events from assessment-events
