+++ Dockerfile
@@ -21,7 +21,7 @@ RUN addgroup --gid 2000 --system appgroup && \
     adduser --uid 2000 --system appuser --gid 2000
 
 WORKDIR /app
-COPY --from=builder --chown=appuser:appgroup /app/build/libs/offender-assessment-events*.jar /app/app.jar
+COPY --from=builder --chown=appuser:appgroup /app/build/libs/offender-assessments-events*.jar /app/app.jar
 COPY --from=builder --chown=appuser:appgroup /app/build/libs/applicationinsights-agent*.jar /app/agent.jar
 COPY --from=builder --chown=appuser:appgroup /app/AI-Agent.xml /app
