+++ helm_deploy/offender-assessment-events/templates/deployment.yaml
@@ -34,7 +34,7 @@ spec:
                         - {{ template "app.name" . }}
                 topologyKey: kubernetes.io/hostname
       containers:
-        - name: offender-assessment-events
+        - name: offender-assessments-events
           image: "{{ .Values.image.repository }}:{{ .Values.image.tag }}"
           ports:
             - containerPort: {{ .Values.image.port }}
