+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessments_events/config/OffenderAssessmentEventsExceptionHandler.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.config
+package uk.gov.justice.digital.hmpps.assessment_events.config
 
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
