+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessments_events/health/HealthInfo.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.health
+package uk.gov.justice.digital.hmpps.assessment_events.health
 
 import org.springframework.boot.actuate.health.Health
 import org.springframework.boot.actuate.health.HealthIndicator
