--- src/main/kotlin/uk/gov/justice/digital/hmpps/offender-assessment-events/OffenderAssessmentEvents.kt
@@ -1,11 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication()
-class OffenderAssessmentEvents
-
-fun main(args: Array<String>) {
-  runApplication<OffenderAssessmentEvents>(*args)
-}
