+++ src/main/resources/application.yml
@@ -4,7 +4,7 @@ info.app:
 
 spring:
   application:
-    name: offender-assessment-events
+    name: offender-assessments-events
   codec:
     max-in-memory-size: 10MB
