+++ src/test/kotlin/uk/gov/justice/digital/hmpps/assessments_events/integration/IntegrationTestBase.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.integration
+package uk.gov.justice.digital.hmpps.assessments_events.integration
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
